Enable ansible ssh pipelining globally
See https://github.com/ansible/ansible/issues/16048#issuecomment-229012509 for why we do this.

The drawback of this is that it is incompatible with sudo's requiretty setting (or su, which always requires a tty). This is because of a quirk of the Python interpreter, which enters interactive mode automatically when you pipe in data from a (pseudo) tty.

However, modern Debian, which we run, comes with requiretty disabled.