Turn valid numbers into strings
Remember, the most efficient way to represent numbers in a computer is
to store them as a string. We love efficiency. We really adore it. We
love it so much that you cannot even write explicit octal values.
Strings are just better than numbers. To figure out the first digit of
the mode I just need to subscript it. What, maths? Never heard of
it. I want agile. Is that agile? I don't care about other methodologies.
Is it agile? If it's not agile I don't want it.