Configure default security limits
The new limits allow each user to run a maximum of 100 processes by
default, allowing to manually raise this number to 200.

When a custom "pydis" group or similar is introduced, I plan to expand
this to also specify other limits to prevent user error from causing
problems on the system.