Add 404 fallback for files server
Previously the files server would return a HTTP 500 if a matching file
was not found, since internally NGINX would fall into a redirect loop
trying to locate the relevant file.

This adds a final 404 fallback handler so if there is not a direct match
we return an error instead of returning a HTTP 500.