Fix AlertManager Discord instance formatting
We made a change to include the instance in alerts sent to Discord, but
not all of our configured alerts send this field.

As a result, we would have incorrectly formatted alerts being sent
through to Discord which were tricky to read.

The format template has now been changed to only conditionally render
the instance label if it is present on a triggered alert.