Add pydis-mtls role for distributing root CA
Adds a new role named pydis-mtls to distribute the mTLS certificate
authority data to all nodes in the inventory.

The defaults are sufficient here and are using the production CA that
will be used for service authentication (tls.pydis.wtf).

Other services can point to the value stored in pydis_mtls_location as
the source of truth for the certificate authority to validate against.