Switch to using designated placeholder IPv4 for originless records
We currently used something like 1.2.3.4 or 1.1.1.1 as placeholder IP
addresses for DNS records where we ran in "originless" mode (the request
is always answered by a Cloudflare Worker or a redirect).

This changes that so we use designated reserved IPv4
addresses (192.0.2.0) to capture that traffic instead, ensuring that in
no circumstance would we leak traffic to an address like 1.1.1.1 or
1.2.3.4 if there was a Cloudflare misconfiguration.

Despite the potential risk vectors here being very small, it's a minor
change and also helps us ensure configuration works correctly in the future.