Don't add limits to LDAP hosts
This prevented the IPA installation from completing, presumably because
it creates a high number of threads/subprocesses/whatever to get the
installation complete.

Regardless, with SELinux and other security tools that ship with the
Rocky security profile we are on. This should be fine. Limits are not a
security risk other than resource exhaustion which is not mission
critical here.

Signed-off-by: Joe Banks <joe@jb3.dev>