Update certificate issuance to group together related certificates
We now can use CSV values to group different (but related) SANs into one
issued certificate.

As an example, when it was migrated in #402, certificates were
configured in such a way that certbot attempted to issue one certificate
for pydis.wtf and another for *.pydis.wtf, which is obviously not
desirable.

This restores previous behaviour to group together certificates served
from the same NGINX vhost, using some Ansible filters to ensure the
`creates` option of the task matches the certbot generated directory.