ManageSieve for Everyone
A spectre is haunting Python Discord -- the spectre of proprietary mail
filtering programs. All the Powers of Big Mail have entered into a holy
alliance to exorcise this spectre: Outlook and GMail, Yahoo and Zoho,
AOL adn iCloud.

Where is the open protocol in opposition that has not been decried as
unprofessional, free and open source by its opponents in power? Where is
the Opposition that has not hurled back the branding reproach of
Internet Standards, against the more advanced opposition protocols, as
well as against its reactionary adversaries?

Two things result from this fact.

I.  Sieve is already standardized as a protocol via the IETF.

II. It is high time that Python Discord users should openly, in the face
of the TLS-encrypted internet, filter their notifications, their love
letters, their mailing lists, and meet this nursery tale of the Spectre
of ManageSieve with a Manifesto of Python Discord Devops itself.

To this end, ManageSieve users of various nationalities have assembled
on jitsi.pydis.wtf, and sketched the following commit, to be published
on the lovelace Python Discord e-mail service.