Treat NGINX responses as UTF-8
Some sites that had non-ASCII characters rendered weirdly because our
NGINX configuration was not returning a default charset and so rendering
was being left up to browser/standards defaults.

This change adds a new config file to /etc/nginx/conf.d/ which forces
responses to be interpreted as/transformed to UTF-8.